Added a .clang-format file so that clang-format will work correctly.
clang-format (http://clang.llvm.org/docs/ClangFormat.html) is a handy tool for
quickly getting all your code in the right style.  One way to use it is to
define a .clang-format file
(see http://clang.llvm.org/docs/ClangFormatStyleOptions.html) in the root
directoy of your project, which will then be used each time you run clang-
format to format your code.  This is a version of that file.  Note that it
should probably be updated so that the ColumnLimit key is something closer to
sane, like 78 or 80.  It is as large as it is right now because I want this
commit to compare relatively cleanly with the mainline.